src/requirements/dev.txt
-r production.txt

django-debug-toolbar==2.0
factory_boy==2.12.0
coverage==4.5.3
codecov==2.0.15
